Move helpers out of apollo dir
