+++ src/app/apollo/resolvers/executeQuery.test.ts
@@ -1,7 +1,7 @@
 import { executeQuery } from './executeQuery';
 
-import { queryBase } from 'app/apollo/helpers/queryBase';
-jest.mock('app/apollo/helpers/queryBase');
+import { queryBase } from '../../helpers/queryBase';
+jest.mock('../../helpers/queryBase');
 
 describe('executeQuery()', () => {
   test('is handled by queryBase() helper', async () => {
