+++ src/app/helpers/executeViaElectron.test.ts
