+++ src/app/helpers/getIpcRenderer.ts
