+++ src/app/helpers/isElectron.test.ts
