+++ src/app/helpers/parseQueryResults.test.ts
