+++ src/app/helpers/parseQueryResults.ts
